+++ packages/openapi_code_builder/example/lib/src/api/petstore.openapi.g.dart
@@ -6,63 +6,33 @@ part of 'petstore.openapi.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Order _$OrderFromJson(Map<String, dynamic> json) {
-  return Order(
-    id: json['id'] as int?,
-    petId: json['petId'] as int?,
-    quantity: json['quantity'] as int?,
-    shipDate: json['shipDate'] == null
-        ? null
-        : DateTime.parse(json['shipDate'] as String),
-    status: _$enumDecodeNullable(_$OrderStatusEnumMap, json['status']),
-    complete: json['complete'] as bool?,
-  );
-}
-
-Map<String, dynamic> _$OrderToJson(Order instance) => <String, dynamic>{
-      'id': instance.id,
-      'petId': instance.petId,
-      'quantity': instance.quantity,
-      'shipDate': instance.shipDate?.toIso8601String(),
-      'status': _$OrderStatusEnumMap[instance.status],
-      'complete': instance.complete,
-    };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
+Order _$OrderFromJson(Map<String, dynamic> json) => Order(
+      id: json['id'] as int?,
+      petId: json['petId'] as int?,
+      quantity: json['quantity'] as int?,
+      shipDate: json['shipDate'] == null
+          ? null
+          : DateTime.parse(json['shipDate'] as String),
+      status: $enumDecodeNullable(_$OrderStatusEnumMap, json['status']),
+      complete: json['complete'] as bool? ?? false,
     );
+
+Map<String, dynamic> _$OrderToJson(Order instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
   }
 
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
-
-K? _$enumDecodeNullable<K, V>(
-  Map<K, V> enumValues,
-  dynamic source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
+  writeNotNull('id', instance.id);
+  writeNotNull('petId', instance.petId);
+  writeNotNull('quantity', instance.quantity);
+  writeNotNull('shipDate', instance.shipDate?.toIso8601String());
+  writeNotNull('status', _$OrderStatusEnumMap[instance.status]);
+  val['complete'] = instance.complete;
+  return val;
 }
 
 const _$OrderStatusEnumMap = {
@@ -71,78 +41,106 @@ const _$OrderStatusEnumMap = {
   OrderStatus.delivered: 'delivered',
 };
 
-Category _$CategoryFromJson(Map<String, dynamic> json) {
-  return Category(
-    id: json['id'] as int?,
-    name: json['name'] as String?,
-  );
+Category _$CategoryFromJson(Map<String, dynamic> json) => Category(
+      id: json['id'] as int?,
+      name: json['name'] as String?,
+    );
+
+Map<String, dynamic> _$CategoryToJson(Category instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('id', instance.id);
+  writeNotNull('name', instance.name);
+  return val;
 }
 
-Map<String, dynamic> _$CategoryToJson(Category instance) => <String, dynamic>{
-      'id': instance.id,
-      'name': instance.name,
-    };
+User _$UserFromJson(Map<String, dynamic> json) => User(
+      id: json['id'] as int?,
+      username: json['username'] as String?,
+      firstName: json['firstName'] as String?,
+      lastName: json['lastName'] as String?,
+      email: json['email'] as String?,
+      password: json['password'] as String?,
+      phone: json['phone'] as String?,
+      userStatus: json['userStatus'] as int?,
+    );
 
-User _$UserFromJson(Map<String, dynamic> json) {
-  return User(
-    id: json['id'] as int?,
-    username: json['username'] as String?,
-    firstName: json['firstName'] as String?,
-    lastName: json['lastName'] as String?,
-    email: json['email'] as String?,
-    password: json['password'] as String?,
-    phone: json['phone'] as String?,
-    userStatus: json['userStatus'] as int?,
-  );
+Map<String, dynamic> _$UserToJson(User instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('id', instance.id);
+  writeNotNull('username', instance.username);
+  writeNotNull('firstName', instance.firstName);
+  writeNotNull('lastName', instance.lastName);
+  writeNotNull('email', instance.email);
+  writeNotNull('password', instance.password);
+  writeNotNull('phone', instance.phone);
+  writeNotNull('userStatus', instance.userStatus);
+  return val;
 }
 
-Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
-      'id': instance.id,
-      'username': instance.username,
-      'firstName': instance.firstName,
-      'lastName': instance.lastName,
-      'email': instance.email,
-      'password': instance.password,
-      'phone': instance.phone,
-      'userStatus': instance.userStatus,
-    };
+Tag _$TagFromJson(Map<String, dynamic> json) => Tag(
+      id: json['id'] as int?,
+      name: json['name'] as String?,
+    );
 
-Tag _$TagFromJson(Map<String, dynamic> json) {
-  return Tag(
-    id: json['id'] as int?,
-    name: json['name'] as String?,
-  );
+Map<String, dynamic> _$TagToJson(Tag instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('id', instance.id);
+  writeNotNull('name', instance.name);
+  return val;
 }
 
-Map<String, dynamic> _$TagToJson(Tag instance) => <String, dynamic>{
-      'id': instance.id,
-      'name': instance.name,
-    };
+Pet _$PetFromJson(Map<String, dynamic> json) => Pet(
+      id: json['id'] as int?,
+      category: json['category'] == null
+          ? null
+          : Category.fromJson(json['category'] as Map<String, dynamic>),
+      name: json['name'] as String,
+      photoUrls:
+          (json['photoUrls'] as List<dynamic>).map((e) => e as String).toList(),
+      tags: (json['tags'] as List<dynamic>?)
+          ?.map((e) => Tag.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      status: $enumDecodeNullable(_$PetStatusEnumMap, json['status']),
+    );
 
-Pet _$PetFromJson(Map<String, dynamic> json) {
-  return Pet(
-    id: json['id'] as int?,
-    category: json['category'] == null
-        ? null
-        : Category.fromJson(json['category'] as Map<String, dynamic>),
-    name: json['name'] as String,
-    photoUrls:
-        (json['photoUrls'] as List<dynamic>).map((e) => e as String).toList(),
-    tags: (json['tags'] as List<dynamic>?)
-        ?.map((e) => Tag.fromJson(e as Map<String, dynamic>))
-        .toList(),
-    status: _$enumDecodeNullable(_$PetStatusEnumMap, json['status']),
-  );
+Map<String, dynamic> _$PetToJson(Pet instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('id', instance.id);
+  writeNotNull('category', instance.category);
+  val['name'] = instance.name;
+  val['photoUrls'] = instance.photoUrls;
+  writeNotNull('tags', instance.tags);
+  writeNotNull('status', _$PetStatusEnumMap[instance.status]);
+  return val;
 }
-
-Map<String, dynamic> _$PetToJson(Pet instance) => <String, dynamic>{
-      'id': instance.id,
-      'category': instance.category,
-      'name': instance.name,
-      'photoUrls': instance.photoUrls,
-      'tags': instance.tags,
-      'status': _$PetStatusEnumMap[instance.status],
-    };
 
 const _$PetStatusEnumMap = {
   PetStatus.available: 'available',
@@ -150,83 +148,68 @@ const _$PetStatusEnumMap = {
   PetStatus.sold: 'sold',
 };
 
-ApiResponse _$ApiResponseFromJson(Map<String, dynamic> json) {
-  return ApiResponse(
-    code: json['code'] as int?,
-    type: json['type'] as String?,
-    message: json['message'] as String?,
-  );
+ApiResponse _$ApiResponseFromJson(Map<String, dynamic> json) => ApiResponse(
+      code: json['code'] as int?,
+      type: json['type'] as String?,
+      message: json['message'] as String?,
+    );
+
+Map<String, dynamic> _$ApiResponseToJson(ApiResponse instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('code', instance.code);
+  writeNotNull('type', instance.type);
+  writeNotNull('message', instance.message);
+  return val;
 }
-
-Map<String, dynamic> _$ApiResponseToJson(ApiResponse instance) =>
-    <String, dynamic>{
-      'code': instance.code,
-      'type': instance.type,
-      'message': instance.message,
-    };
-
-FindPetsByStatusResponseBody200 _$FindPetsByStatusResponseBody200FromJson(
-    Map<String, dynamic> json) {
-  return FindPetsByStatusResponseBody200();
-}
-
-Map<String, dynamic> _$FindPetsByStatusResponseBody200ToJson(
-        FindPetsByStatusResponseBody200 instance) =>
-    <String, dynamic>{};
-
-FindPetsByTagsResponseBody200 _$FindPetsByTagsResponseBody200FromJson(
-    Map<String, dynamic> json) {
-  return FindPetsByTagsResponseBody200();
-}
-
-Map<String, dynamic> _$FindPetsByTagsResponseBody200ToJson(
-        FindPetsByTagsResponseBody200 instance) =>
-    <String, dynamic>{};
 
 UpdatePetWithFormSchema _$UpdatePetWithFormSchemaFromJson(
-    Map<String, dynamic> json) {
-  return UpdatePetWithFormSchema(
-    name: json['name'] as String?,
-    status: json['status'] as String?,
-  );
-}
+        Map<String, dynamic> json) =>
+    UpdatePetWithFormSchema(
+      name: json['name'] as String?,
+      status: json['status'] as String?,
+    );
 
 Map<String, dynamic> _$UpdatePetWithFormSchemaToJson(
-        UpdatePetWithFormSchema instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'status': instance.status,
-    };
+    UpdatePetWithFormSchema instance) {
+  final val = <String, dynamic>{};
 
-UploadFileSchema _$UploadFileSchemaFromJson(Map<String, dynamic> json) {
-  return UploadFileSchema();
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('name', instance.name);
+  writeNotNull('status', instance.status);
+  return val;
 }
-
-Map<String, dynamic> _$UploadFileSchemaToJson(UploadFileSchema instance) =>
-    <String, dynamic>{};
 
 GetInventoryResponseBody200 _$GetInventoryResponseBody200FromJson(
-    Map<String, dynamic> json) {
-  return GetInventoryResponseBody200();
-}
+        Map<String, dynamic> json) =>
+    GetInventoryResponseBody200();
 
 Map<String, dynamic> _$GetInventoryResponseBody200ToJson(
         GetInventoryResponseBody200 instance) =>
     <String, dynamic>{};
 
 CreateUsersWithArrayInputSchema _$CreateUsersWithArrayInputSchemaFromJson(
-    Map<String, dynamic> json) {
-  return CreateUsersWithArrayInputSchema();
-}
+        Map<String, dynamic> json) =>
+    CreateUsersWithArrayInputSchema();
 
 Map<String, dynamic> _$CreateUsersWithArrayInputSchemaToJson(
         CreateUsersWithArrayInputSchema instance) =>
     <String, dynamic>{};
 
 LoginUserResponseBody200 _$LoginUserResponseBody200FromJson(
-    Map<String, dynamic> json) {
-  return LoginUserResponseBody200();
-}
+        Map<String, dynamic> json) =>
+    LoginUserResponseBody200();
 
 Map<String, dynamic> _$LoginUserResponseBody200ToJson(
         LoginUserResponseBody200 instance) =>