+++ packages/openapi_code_builder/example/lib/src/api/testapi.openapi.g.dart
@@ -6,46 +6,57 @@ part of 'testapi.openapi.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-RegisterRequest _$RegisterRequestFromJson(Map<String, dynamic> json) {
-  return RegisterRequest(
-    email: json['email'] as String,
-  );
-}
+RegisterRequest _$RegisterRequestFromJson(Map<String, dynamic> json) =>
+    RegisterRequest(
+      email: json['email'] as String,
+    );
 
 Map<String, dynamic> _$RegisterRequestToJson(RegisterRequest instance) =>
     <String, dynamic>{
       'email': instance.email,
     };
 
-HelloRequest _$HelloRequestFromJson(Map<String, dynamic> json) {
-  return HelloRequest(
-    salutation: json['salutation'] as String?,
-  );
+HelloRequest _$HelloRequestFromJson(Map<String, dynamic> json) => HelloRequest(
+      salutation: json['salutation'] as String?,
+    );
+
+Map<String, dynamic> _$HelloRequestToJson(HelloRequest instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('salutation', instance.salutation);
+  return val;
 }
 
-Map<String, dynamic> _$HelloRequestToJson(HelloRequest instance) =>
-    <String, dynamic>{
-      'salutation': instance.salutation,
-    };
+HelloResponse _$HelloResponseFromJson(Map<String, dynamic> json) =>
+    HelloResponse(
+      message: json['message'] as String?,
+    );
 
-HelloResponse _$HelloResponseFromJson(Map<String, dynamic> json) {
-  return HelloResponse(
-    message: json['message'] as String?,
-  );
+Map<String, dynamic> _$HelloResponseToJson(HelloResponse instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('message', instance.message);
+  return val;
 }
-
-Map<String, dynamic> _$HelloResponseToJson(HelloResponse instance) =>
-    <String, dynamic>{
-      'message': instance.message,
-    };
 
 UuidExampleMessageIdGetResponseBody200
     _$UuidExampleMessageIdGetResponseBody200FromJson(
-        Map<String, dynamic> json) {
-  return UuidExampleMessageIdGetResponseBody200(
-    id: const ApiUuidJsonConverter().fromJson(json['id'] as String),
-  );
-}
+            Map<String, dynamic> json) =>
+        UuidExampleMessageIdGetResponseBody200(
+          id: const ApiUuidJsonConverter().fromJson(json['id'] as String),
+        );
 
 Map<String, dynamic> _$UuidExampleMessageIdGetResponseBody200ToJson(
         UuidExampleMessageIdGetResponseBody200 instance) =>